[Refactor] #27 StickerView 리펙토링
미디어 종류 별로 분기.
